requirements.txt
Flask
requests
beautifulsoup4
pandas
apscheduler
openpyxl
python-dotenv
